requirements.txt
matplotlib==3.3.1
seaborn==0.11.0
numpy==1.18.5
pandas==1.1.2
plotly==4.12.0
streamlit==0.79.0
Pillow==8.2.0
scikit_learn==0.24.1
